Create First Lab
The current lab is the first lab of the fourth semester which the teacher has given me the number 8. 
And the current lab is also the 8th and  all the rules and the conditions of the lab is also in the folder which is in my desctop and there is the number 8. The file is also in word form.